Refactor Google API access for reuse
