+++ app/api/bafko/route.ts
@@ -1,38 +1,9 @@
 'use server';
 import { NextResponse } from 'next/server';
-import {google, sheets_v4} from 'googleapis';
-import {GaxiosResponse} from 'gaxios';
+import {getSheetValues, SheetValues} from '@/util/GoogleApi';
 
-function getJwt() {
-    return new google.auth.JWT(
-        process.env.SERVICE_CLIENT_EMAIL, undefined, process.env.SERVICE_CLIENT_KEY,
-        ['https://www.googleapis.com/auth/spreadsheets']
-    );
-}
-
-type SheetContent =  GaxiosResponse<sheets_v4.Schema$ValueRange> | null | undefined
-
-async function getSheet(): Promise<SheetContent> {
-    const sheets = google.sheets({version: 'v4'});
-    return new Promise((resolve, reject) => {
-        sheets.spreadsheets.values.get({
-            spreadsheetId: process.env.BAFKO_SHEET_ID,
-            range: 'A3:C',
-            auth: getJwt(),
-            key: process.env.SHEETS_API_KEY,
-        }, (err, res) => {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(res);
-            }
-        });
-    });
-}
-
-const getBafkos = async (): Promise<GaxiosResponse<sheets_v4.Schema$ValueRange>['data']['values'] | null> => {
-    const sheetData = await getSheet();
-    return sheetData?.data.values;
+const getBafkos = async (): Promise<SheetValues | null> => {
+    return getSheetValues(process.env.BAFKO_SHEET_ID!, 'A3:C');
 };
 
 export async function GET() {
